requirements.txt
asgiref==3.5.0
Django==3.1.8
django-cors-headers==3.11.0
django-userforeignkey==0.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
PyJWT==2.3.0
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
